+++ tests/ekobox/test_genpdf_load.py
@@ -1,5 +1,3 @@
-import numpy as np
-
 from ekobox import genpdf
 
 
@@ -11,11 +9,5 @@ def test_load_info(fake_ct14):
 
 
 def test_load_data_ct14(fake_ct14):
-    blocks = genpdf.load.load_blocks_from_file(fake_ct14, 0)[1]
-    assert len(blocks) == 1
-    b0 = blocks[0]
-    assert isinstance(b0, dict)
-    assert sorted(b0.keys()) == sorted(["pids", "xgrid", "mu2grid", "data"])
-    assert sorted(b0["pids"]) == sorted([-3, -2, -1, 21, 1, 2, 3])
-    assert len(b0["data"].T) == 7
-    np.testing.assert_allclose(b0["xgrid"][0], 1e-9)
+    f = genpdf.load.load_blocks_from_file(fake_ct14, 0)
+    assert len(f.blocks) == 1